pyproject.toml
[tool.poetry]
name = "strategyexplainer"
version = "0.1.0"
description = ""
authors = ["Michael Portale <mvportale@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
yfinance = "^0.2.63"
pandas = "^2.3.0"
matplotlib = "^3.10.3"
quantstats = "^0.0.64"
python-dotenv = "^1.1.0"
streamlit = "^1.28.0"
plotly = "^5.17.0"
openai = "^1.0.0"
scipy = "^1.11.0"
pyyaml = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"